--- ktlint-baseline/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-    id("ktlint-publication-library")
-}
-
-dependencies {
-    implementation(projects.ktlintLogger)
-    implementation(projects.ktlintRuleEngineCore)
-    implementation(projects.ktlintCliReporterCore)
-}
