--- ktlint-baseline/gradle.properties
@@ -1,2 +1,0 @@
-POM_NAME=ktlint-baseline
-POM_ARTIFACT_ID=ktlint-baseline
