+++ ktlint-cli-reporter-baseline/build.gradle.kts
@@ -3,6 +3,8 @@ plugins {
 }
 
 dependencies {
+    implementation(projects.ktlintLogger)
+    implementation(projects.ktlintRuleEngineCore)
     implementation(projects.ktlintCliReporterCore)
 
     testImplementation(projects.ktlintTest)
